third_party/blink/web_tests/platform/generic/http/tests/inspector-protocol/storage/cookies-expected.txt

Tests that cookies are read and written.

Running test: testPageContext

Initial: 0 cookies
Cookies as seen on server: "HTTP_COOKIE: <not set>\n"

Post-fetch: 1 cookies
name: name, value: value, domain: 127.0.0.1, path: /inspector-protocol/network/resources, session
Cookies as seen on server: "HTTP_COOKIE: name=value\n"

Post-set: 3 cookies
name: foo, value: second bar2, domain: 127.0.0.1, path: /, session
name: foo2, value: bar1, domain: 127.0.0.1, path: /, session
name: name, value: value, domain: 127.0.0.1, path: /inspector-protocol/network/resources, session
Cookies as seen on server: "HTTP_COOKIE: name=value; foo=second bar2; foo2=bar1\n"

Post-clear: 0 cookies
Cookies as seen on server: "HTTP_COOKIE: <not set>\n"

Running test: testInvalidContext
{
    error : {
        code : -32602
        message : Failed to find browser context for id invalid
    }
    id : <number>
    sessionId : <string>
}
